sync function to find mdFiles
