--- MDLinks-Async/mdLinksAsync.js
@@ -1,17 +1,0 @@
-const fs = require('fs');
-const path = require('path');
-const functionFiles = require('./find-FilesAsync.js')
-//const functionLinks = require('./findLinks.js')
-
-const route = process.argv[2];
-
-  const files = functionFiles.findFiles(route) 
-  //.then( data => { console.log(data)})
-  //console.log(files)
-  //const links = functionLinks.findLinks(files);
-
-
-
-
-//node mdLinks 'C:/Users/Colette/Desktop/Labo/CDMX011-md-links'
-//node mdLinks 'C:/Users/Colette/Desktop/Labo/CDMX011-md-links/README.md'
